30ex SDL3版最终代码

+++ src/Object.h
@@ -1,6 +1,6 @@
 #ifndef OBJECT_H
 #define OBJECT_H
-#include <SDL.h>
+#include <SDL3/SDL.h>
 
 enum class ItemType{
     Life,
@@ -11,8 +11,8 @@ enum class ItemType{
 struct Player{
     SDL_Texture* texture = nullptr;
     SDL_FPoint position = {0, 0};
-    int width = 0;
-    int height = 0;
+    float width = 0;
+    float height = 0;
     int speed = 300;
     int currentHealth = 3;
     int maxHealth = 3;
@@ -23,8 +23,8 @@ struct Player{
 struct Enemy{
     SDL_Texture* texture = nullptr;
     SDL_FPoint position = {0, 0};
-    int width = 0;
-    int height = 0;
+    float width = 0;
+    float height = 0;
     int speed = 150;
     int currentHealth = 2;
     Uint32 coolDown = 2000;
@@ -34,8 +34,8 @@ struct Enemy{
 struct ProjectilePlayer{
     SDL_Texture* texture = nullptr;
     SDL_FPoint position = {0, 0};
-    int width = 0;
-    int height = 0;
+    float width = 0;
+    float height = 0;
     int speed = 600;
     int damage = 1;
 };
@@ -44,8 +44,8 @@ struct ProjectileEnemy{
     SDL_Texture* texture = nullptr;
     SDL_FPoint position = {0, 0};
     SDL_FPoint direction = {0, 0};
-    int width = 0;
-    int height = 0;
+    float width = 0;
+    float height = 0;
     int speed = 400;
     int damage = 1;
 };
@@ -53,8 +53,8 @@ struct ProjectileEnemy{
 struct Explosion{
     SDL_Texture* texture = nullptr;
     SDL_FPoint position = {0, 0};
-    int width = 0;
-    int height = 0;
+    float width = 0;
+    float height = 0;
     int currentFrame = 0;
     int totlaFrame = 0;
     Uint32 startTime = 0;
@@ -65,8 +65,8 @@ struct Item{
     SDL_Texture* texture = nullptr;
     SDL_FPoint position = {0, 0};
     SDL_FPoint direction = {0, 0};
-    int width = 0;
-    int height = 0;
+    float width = 0;
+    float height = 0;
     int speed = 200;
     int bounceCount = 3;
     ItemType type = ItemType::Life;
@@ -76,8 +76,8 @@ struct Background{
     SDL_Texture* texture = nullptr;
     SDL_FPoint position = {0, 0};
     float offset = 0;
-    int width = 0;
-    int height = 0;
+    float width = 0;
+    float height = 0;
     int speed = 30;
 };
